% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceR.R
\name{AggStructure}
\alias{AggStructure}
\title{AggStructure}
\usage{
AggStructure(x)
}
\arguments{
\item{x}{aggregated accounts as a data frame with columns
"account", "amount". Account hierarchy must be seperated by a ":"
in the account names according to ledger convention}
}
\value{
A data frame containing one side of the balance sheet
   already in the correct order.
}
\description{
Create the aggregate structure from already aggregated balance sheet
}
\details{
function only in beta state...
}
