% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apg}
\alias{apg}
\title{Accelerated proximal gradient method}
\usage{
apg(grad_ptr, prox_ptr, loss_opts, prox_opts, dim1, dim2, max_it, eps, alpha,
  beta, accel)
}
\arguments{
\item{grad_ptr}{Pointer to gradient function}

\item{prox_ptr}{Pointer to prox function}

\item{loss_opts}{List of options for loss (input data, tuning params, etc.)}

\item{max_it}{Maximum number of iterations}

\item{eps}{Convergence tolerance}

\item{beta}{Backtracking line search parameter}

\item{loss_ptr}{Pointer to loss function}

\item{dim}{Dimension}
}
\value{
Optimal value
}
\description{
Accelerated proximal gradient method
}
