% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apg_warmstart}
\alias{apg_warmstart}
\title{Accelerated proximal gradient method}
\usage{
apg_warmstart(
  grad_ptr,
  prox_ptr,
  loss_opts,
  prox_opts,
  lams,
  x,
  max_it,
  eps,
  alpha,
  beta,
  accel,
  verbose
)
}
\arguments{
\item{grad_ptr}{Pointer to gradient function}

\item{prox_ptr}{Pointer to prox function}

\item{loss_opts}{List of options for loss (input data, tuning params, etc.)}

\item{prox_opts}{List of options for prox (regularization parameter)}

\item{lams}{Vector of hyperparameters to use}

\item{x}{Initial value}

\item{max_it}{Maximum number of iterations}

\item{eps}{Convergence tolerance}

\item{beta}{Backtracking line search parameter}

\item{verbose}{How much information to print}
}
\value{
Optimal value
}
\description{
Accelerated proximal gradient method
}
