% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancer.R, R/main_funcs.R
\docType{package}
\name{balancer}
\alias{balancer}
\alias{balancer-package}
\alias{balancer}
\title{balancer: A package for treatment effect heterogeneity and missing data in observational studies through matrix-constrained balancing weights}
\usage{
balancer(X, trt, Z = NULL, V = NULL, type = c("att", "subgrp",
  "subgrp_multi", "multimatch", "missing", "hte"), link = c("logit", "linear",
  "pos-linear", "pos-enet", "posenet"), regularizer = c(NULL, "l1", "grpl1",
  "l2", "ridge", "linf", "nuc", "l1_all", "l1_nuc"), hyperparam, interact = F,
  Q = NULL, kernel = NULL, kern_param = 1, normalized = TRUE,
  opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of subgroup indicators or observed indicators}

\item{V}{Group level covariates}

\item{type}{Find balancing weights for ATT, subgroup ATTs,
subgroup ATTs with multilevel p-score, multilevel observational studies,
ATT with missing outcomes, and heterogeneous effects}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{hyperparam}{Regularization hyperparameter}

\item{interact}{Whether to interact group and individual level covariates}

\item{Q}{m x m matrix to tie together ridge penalty, default: NULL,
if TRUE, use covariance of treated groups}

\item{kernel}{What kernel to use, default NULL}

\item{kern_param}{Hyperparameter for kernel}

\item{normalized}{Whether to normalize the weights}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}
  \item{alpha }{Elastic net parameter}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
balancer: A package for treatment effect heterogeneity and missing data in observational studies through matrix-constrained balancing weights

Find Balancing weights by solving the dual optimization problem
}
