% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{balancer_att}
\alias{balancer_att}
\title{Balancing weights for ATT (in subgroups)}
\usage{
balancer_att(X, trt, y = NULL, weightfunc, weightfunc_ptr, proxfunc,
  balancefunc, lambda = NULL, nlambda = 20, lambda.min.ratio = 0.001,
  ipw_weights = NULL, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{y}{Vector of outcomes to estimate effect(s). If NULL then only return weights}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function}

\item{balancefunc}{Balance criterion measure}

\item{lambda}{Regularization hyper parameter}

\item{nlambda}{Number of hyperparameters to consider}

\item{lambda.min.ratio}{Smallest value of hyperparam to consider, as proportion of smallest
value that gives the reference weights}

\item{ipw_weights}{Separately estimated IPW weights to measure dispersion against, default is NULL}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}

\item{ridge}{Whether to use L2 penalty in dual}

\item{Q}{m x m matrix to tie together ridge penalty, default: NULL,
if TRUE, use covariance of treated groups}

\item{kernel}{What kernel to use, default NULL}

\item{kern_param}{Hyperparameter for kernel}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for ATT (in subgroups)
}
