% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{balancer_att_cv}
\alias{balancer_att_cv}
\title{Balancing weights for ATT (in subgroups)}
\usage{
balancer_att_cv(
  X,
  trt,
  k,
  y = NULL,
  weightfunc,
  weightfunc_ptr,
  proxfunc,
  balancefunc,
  lambda = NULL,
  nlambda = 20,
  lambda.min.ratio = 0.001,
  ipw_weights = NULL,
  mu0 = NULL,
  opts = list(),
  prox_opts = list()
)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{k}{Number of folds}

\item{y}{Vector of outcomes to estimate effect(s). If NULL then only return weights}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function}

\item{balancefunc}{Balance criterion measure}

\item{lambda}{Vector of regularization hyperparameters to consider, if NULL then uses
a data-generated list}

\item{nlambda}{Number of hyperparameters to consider}

\item{lambda.min.ratio}{Smallest value of hyperparam to consider, as proportion of smallest
value that gives the reference weights}

\item{ipw_weights}{Separately estimated IPW weights to measure dispersion against, default is NULL}

\item{mu0}{Optional estimates of the potential outcome under control, default is NULL}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}

\item{prox_opts}{List of additional arguments for prox}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for ATT (in subgroups)
}
