% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{balancer_int}
\alias{balancer_int}
\title{Helper function to fit the dual for general odds function and prox}
\usage{
balancer_int(X, trt, Z, conjprime, proxfunc, hyperparam, normalized = TRUE)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of subgroup indicators}

\item{conjprime}{Derivative of convex conjugate of dispersion function}

\item{proxfunc}{Prox operator of regularization function}

\item{hyperparam}{Regularization hyper parameter}

\item{normalized}{Whether to fit normalized weights, default: True}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Helper function to fit the dual for general odds function and prox
}
