% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_funcs.R
\name{balancer_missing}
\alias{balancer_missing}
\title{Balancing weights for missing outcomes}
\usage{
balancer_missing(X, trt, R, weightfunc, weightfunc_ptr, proxfunc,
  hyperparam, ridge, Q, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{R}{vector of missingness indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function}

\item{hyperparam}{Regularization hyper parameter}

\item{ridge}{Whether to use L2 penalty in dual}

\item{Q}{m x m matrix to tie together ridge penalty, default: NULL,
if TRUE, use covariance of treated groups}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for missing outcomes
}
