% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_old.R
\name{balancer_missing_old}
\alias{balancer_missing_old}
\title{Helper function to fit the dual for general odds function and prox
Estimates ATT with missing outcomes}
\usage{
balancer_missing_old(X, trt, R, weightfunc, proxfunc, hyperparam,
  opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{R}{vector of missingness indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{proxfunc}{Prox operator of regularization function}

\item{hyperparam}{Regularization hyper parameter}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Helper function to fit the dual for general odds function and prox
Estimates ATT with missing outcomes
}
