% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{balancer_multi}
\alias{balancer_multi}
\title{Balancing weights for ATT (with hierarchical structure)}
\usage{
balancer_multi(
  X,
  V = NULL,
  trt,
  Z,
  weightfunc,
  weightfunc_ptr,
  proxfunc,
  balancefunc,
  lambda,
  nlambda,
  lambda.min.ratio,
  ipw_weights,
  interact = F,
  opts = list(),
  prox_opts
)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{V}{n x k matrix of group level covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of hierarchical factor indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function}

\item{balancefunc}{Balance criterion measure}

\item{lambda}{Regularization hyper parameter}

\item{nlambda}{Number of hyperparameters to consider}

\item{lambda.min.ratio}{Smallest value of hyperparam to consider, as proportion of smallest
value that gives the reference weights}

\item{ipw_weights}{Separately estimated IPW weights to measure dispersion against, default is NULL}

\item{interact}{Whether to interact group and individual level characterisitics}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}

\item{prox_opts}{List of additional arguments for prox}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for ATT (with hierarchical structure)
}
