% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_old.R
\name{balancer_old}
\alias{balancer_old}
\title{Find Balancing weights by solving the dual optimization problem}
\usage{
balancer_old(X, trt, Z = NULL, type = c("att", "subgrp", "missing",
  "hte"), link = c("logit", "linear", "pos-linear", "pos-enet",
  "posenet"), regularizer = c(NULL, "l1", "grpl1", "l2", "ridge", "linf",
  "nuc"), hyperparam, normalized = TRUE, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of subgroup indicators or observed indicators}

\item{type}{Find balancing weights for ATT, subgroup ATTs,
ATT with missing outcomes, and heterogeneouts effects}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{hyperparam}{Regularization hyperparameter}

\item{normalized}{Whether to normalize the weights}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}
  \item{alpha }{Elastic net parameter}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Find Balancing weights by solving the dual optimization problem
}
