% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{balancer_subgrp}
\alias{balancer_subgrp}
\title{Helper function to fit the dual for general odds function and prox
estimates heterogeneous treatment effects}
\usage{
balancer_subgrp(X, trt, Z = NULL, weightfunc, proxfunc, hyperparam,
  normalized = TRUE, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of subgroup indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{proxfunc}{Prox operator of regularization function}

\item{hyperparam}{Regularization hyper parameter}

\item{normalized}{Whether to fit normalized weights, default: True}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Helper function to fit the dual for general odds function and prox
estimates heterogeneous treatment effects
}
