% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{balancing_grad_kern}
\alias{balancing_grad_kern}
\title{Generic balancing loss gradient with kernels}
\usage{
balancing_grad_kern(theta, opts)
}
\description{
Generic balancing loss gradient with kernels
}
