\name{BlcDF}
\alias{BlcDF}
\docType{data}
\title{
Dummy data
}
\description{
Dummy data for BalanceR
}
\usage{data("BlcDF")}
\format{
  A data frame with 2856 observations on the following 6 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Group}}{a character vector}
    \item{\code{Sex}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Educ}}{a numeric vector}
    \item{\code{Marriage}}{a numeric vector}
  }
}
\examples{
data(BlcDF)
}
\keyword{datasets}
