% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate sample to target}
\usage{
calibrate(
  formula,
  target_count,
  data,
  order = NULL,
  lambda = 1,
  prob_weights = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{sample_count ~ covariates}

\item{target_count}{Number in cell in target}

\item{data}{Dataframe with covariate information, sample and target counts}

\item{order}{What order interactions to balance}

\item{lambda}{Regularization hyperparamter}

\item{prob_weights}{Optional sampling weights to include}
}
\description{
Calibrate sample to target
}
