% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_approx.R
\name{calibrate_approx}
\alias{calibrate_approx}
\title{Approximate calibration with categorical covariates}
\usage{
calibrate_approx(X, target, lambda, lowlim = 0, uplim = 1,
  verbose = TRUE, eps_abs = 1e-05, eps_rel = 1e-05, ...)
}
\description{
Approximate calibration with categorical covariates
}
