% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment.R
\name{check_args_treatment}
\alias{check_args_treatment}
\title{Check that data is in right shape and hyparameters are feasible}
\usage{
check_args_treatment(
  X0,
  Xtau,
  target,
  S,
  Z,
  pscores,
  X0s,
  Xtaus,
  nj,
  lambda,
  lowlim,
  uplim,
  data_in
)
}
\arguments{
\item{X0}{n x d0 matrix of covariates}

\item{Xtau}{n x dtau matrix of covariates}

\item{target}{Vector of population means to re-weight to}

\item{S}{Vector of group indicators with J levels}

\item{Z}{Vector of treatment indicators}

\item{pscores}{Vector of propensity scores}

\item{X0s}{List of J n x d0 matrices of covariates split by group}

\item{Xtaus}{List of J n x dtau matrices of covariates split by group}

\item{nj}{Number of units in jth site}

\item{lambda}{Regularization hyper parameter}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{data_in}{Optional list containing pre-computed objective matrix/vector and constraints (without regularization term)}
}
\description{
Check that data is in right shape and hyparameters are feasible
}
