% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_weights.R
\name{check_data_cluster}
\alias{check_data_cluster}
\title{Check that data is in right shape and hyparameters are feasible}
\usage{
check_data_cluster(ind_covs, clus_covs, trt, lambda, lowlim, uplim)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{clus_covs}{n x d2 matrix of covariates for clusters}

\item{trt}{n vector of treatment assignment}

\item{lambda}{Regularization hyper parameter}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}
}
\description{
Check that data is in right shape and hyparameters are feasible
}
