% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_qp.R
\name{check_data_multi}
\alias{check_data_multi}
\title{Check that data is in right shape and hyparameters are feasible}
\usage{
check_data_multi(X, trt, Z, Xz, lambda, lowlim, uplim)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{Z}{Vector of group indicators with J levels}

\item{Xz}{list of J n x d matrices of covariates split by group}

\item{lambda}{Regularization hyper parameter}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}
}
\description{
Check that data is in right shape and hyparameters are feasible
}
