% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_rct.R
\name{check_data_rct}
\alias{check_data_rct}
\title{Check that data is in right shape and hyparameters are feasible}
\usage{
check_data_rct(X, trt, target, Z, Xz, lambda, lowlim, uplim, data_in)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment assignments}

\item{target}{Vector of population means to re-weight to}

\item{Z}{Vector of group indicators with J levels}

\item{Xz}{list of J n x d matrices of covariates split by group}

\item{lambda}{Regularization hyper parameter}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{data_in}{Optional list containing pre-computed objective matrix/vector and constraints (without regularization term)}

\item{verbose}{Whether to show messages, default T}

\item{return_data}{Whether to return the objective matrix and vector and constraints, default T}
}
\description{
Check that data is in right shape and hyparameters are feasible
}
