% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_qp.R
\name{check_data_surv}
\alias{check_data_surv}
\title{Check that data is in right shape, hyparameters are feasible, and basis fns are valid}
\usage{
check_data_surv(B_X, trt, times, events, t, lambda, lowlim = 1, uplim = NULL)
}
\arguments{
\item{B_X}{n x k basis matrix of covariates (X)}

\item{trt}{Vector of treatment assignments}

\item{times}{Vector of event/censoring times (see events for if event or censoring time)}

\item{events}{Vector of boolean censoring indicators (whether individual was censored)}

\item{t}{Time}

\item{lambda}{Regularization hyper parameter}

\item{lowlim}{Lower limit on weights, default 1}

\item{uplim}{Upper limit on weights, default NULL}
}
\description{
Check that data is in right shape, hyparameters are feasible, and basis fns are valid
}
