% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_weights.R
\name{cluster_weights}
\alias{cluster_weights}
\title{Re-weight groups to target population means}
\usage{
cluster_weights(
  ind_covs,
  clus_covs,
  trt,
  clusters,
  lambda = 0,
  icc = 0,
  lowlim = 0,
  uplim = 1,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{clus_covs}{n x d2 matrix of covariates for clusters}

\item{trt}{n vector of treatment assignment}

\item{clusters}{n vector of cluster assignments}

\item{lambda}{Regularization hyper parameter, default 0}

\item{icc}{Intraclass correlation coefficient for regularization}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated primal weights}
         \item{ind_imbalance }{Imbalance in individual-level covariates covariates}
         \item{clus_imbalance }{Imbalance in cluster-level covariates covariates}
}
}
\description{
Re-weight groups to target population means
}
