% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment_kernel.R
\name{compute_block_diag_kernel}
\alias{compute_block_diag_kernel}
\title{Compute block diagonal kernel matrix}
\usage{
compute_block_diag_kernel(Xs, kernel, gc)
}
\arguments{
\item{Xs}{List of covariate matrices by site}

\item{kernel}{Kernel to use}

\item{gc}{boolean indicating whether to garbage collect between operations}
}
\description{
Compute block diagonal kernel matrix
}
