% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_weights.R
\name{compute_cluster_se}
\alias{compute_cluster_se}
\title{Compute point estimate and standard error with clustered weights}
\usage{
compute_cluster_se(y, wts, trt, clusters, m1hat, m0hat)
}
\arguments{
\item{y}{Vector of outcomes}

\item{wts}{Vector of weight}

\item{trt}{Vector of treatment assignments}

\item{clusters}{Vector of cluster assignments}

\item{m1hat}{Vector of model predictions of E[Y(1) | covariates]}

\item{m0hat}{Vector of model predictions of E[Y(0) | covariates]}
}
\value{
Data.Frame with the point estimate and standard error
}
\description{
Compute point estimate and standard error with clustered weights
}
