% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_gram}
\alias{compute_gram}
\title{Compute a gram matrix between X and Y
for a given kernel}
\usage{
compute_gram(X, Y, kernel, param)
}
\arguments{
\item{X, }{Y, matrices}

\item{kernel}{Name of kernel}

\item{param}{Hyper parameter for the kernel}
}
\description{
Compute a gram matrix between X and Y
for a given kernel
}
