% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_approx.R
\name{create_calibrate_constraints}
\alias{create_calibrate_constraints}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_calibrate_constraints(X, target, lambda, lowlim, uplim, verbose)
}
\arguments{
\item{X}{Matrix of factors'}

\item{target}{Vector of population means to re-weight to}

\item{lambda}{Balance constraint level}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
