% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R, R/survival_qp.R
\name{create_constraints}
\alias{create_constraints}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints(
  n,
  trt,
  times,
  events,
  t,
  lowlim = 1,
  uplim = NULL,
  verbose = TRUE
)

create_constraints(
  n,
  trt,
  times,
  events,
  t,
  lowlim = 1,
  uplim = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n}{Number of individuals}

\item{trt}{Vector of treatment assignments}

\item{times}{Vector of event/censoring times (see events for if event or censoring time)}

\item{events}{Vector of boolean censoring indicators (whether individual was censored)}

\item{t}{Time}

\item{lowlim}{Lower limit on weights, default 1}

\item{uplim}{Upper limit on weights, default NULL}

\item{verbose}{Whether to show messages, default T}

\item{Xz}{list of J n x d matrices of covariates split by group}

\item{target}{Vector of population means to re-weight to}

\item{Z}{Vector of group indicators}
}
\value{
A, l, and u

A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u

Create the constraints for QP: l <= Ax <= u
}
