% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_weights.R
\name{create_constraints_cluster}
\alias{create_constraints_cluster}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints_cluster(ind_covs, clus_covs, trt, lowlim, uplim,
  verbose)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{clus_covs}{n x d2 matrix of covariates for clusters}

\item{trt}{n vector of treatment assignment}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}

\item{verbose}{Whether to show messages}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
