% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_l2.R
\name{create_constraints_l2}
\alias{create_constraints_l2}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints_l2(
  Xz,
  targetz,
  target_propz,
  lowlim,
  uplim,
  exact_global,
  verbose
)
}
\arguments{
\item{Xz}{length J list of n_j x d matrix of covariates for each group j}

\item{targetz}{length J list of d-dimensional vectors of targets for each group j}

\item{target_propz}{J-dimensional vector of group proportions in the target population}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}

\item{exact_global}{Boolean whether to include an exact global constraint}

\item{verbose}{Boolean whether to display progress}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
