% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mundlak.R
\name{create_constraints_mundlak}
\alias{create_constraints_mundlak}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints_mundlak(
  ind_covs,
  interact_covs,
  trt,
  Z,
  lowlim,
  uplim,
  exact_global,
  verbose
)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{interact_covs}{n x d2 matrix of interactions between individual and cluster covariates}

\item{trt}{Vector of treatment assignments}

\item{Z}{Vector of group indicators with J levels}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}

\item{exact_global}{Boolean whether to include an exact global constraint}

\item{verbose}{Boolean whether to display progress}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
