% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_rct.R
\name{create_constraints_rct}
\alias{create_constraints_rct}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints_rct(
  Xz,
  trt,
  target,
  Z,
  lowlim,
  uplim,
  exact_global,
  constrain_trt_prop,
  verbose
)
}
\arguments{
\item{Xz}{list of J n x d matrices of covariates split by group}

\item{target}{Vector of population means to re-weight to}

\item{Z}{Vector of group indicators}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
