% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{create_folds}
\alias{create_folds}
\title{Partition indices into k folds}
\usage{
create_folds(n, k)
}
\arguments{
\item{n}{Number of data points}

\item{k}{Number of folds}
}
\value{
List of fold indices
}
\description{
Partition indices into k folds
}
