% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{create_I0_matrix}
\alias{create_I0_matrix}
\title{Create diagonal regularization matrix}
\usage{
create_I0_matrix(Xz, scale_sample_size, n, aux_dim)
}
\arguments{
\item{Xz}{list of J n x d matrices of covariates split by group}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default T}

\item{n}{Total number of units}

\item{aux_dim}{Dimension of auxiliary weights}
}
\description{
Create diagonal regularization matrix
}
