% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_qp.R
\name{create_I0_matrix_surv}
\alias{create_I0_matrix_surv}
\title{Create diagonal regularization matrix}
\usage{
create_I0_matrix_surv(n, lambda)
}
\arguments{
\item{n}{Total number of units}

\item{lambda}{Regularization hyperparameter, default 0}
}
\value{
I0 matrix
}
\description{
Create diagonal regularization matrix
}
