% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment_kernel.R
\name{create_I0_matrix_treatment_kernel}
\alias{create_I0_matrix_treatment_kernel}
\title{Create diagonal regularization matrix}
\usage{
create_I0_matrix_treatment_kernel(
  pro_trt_split,
  pro_ctr_split,
  scale_sample_size,
  nj,
  n
)
}
\arguments{
\item{pro_trt_split}{List of J vectors of treatment propensity multipliers}

\item{pro_ctr_split}{List of J vectors of control propensity multipliers}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default T}

\item{nj}{Number of units in each group}

\item{n}{Total number of units}
}
\description{
Create diagonal regularization matrix
}
