% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{create_P_matrix}
\alias{create_P_matrix}
\title{Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_P_matrix(n, aux_dim)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{Z}{Vector of group indicators}
}
\value{
P matrix
}
\description{
Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x
}
