% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_weights.R
\name{create_P_matrix_cluster}
\alias{create_P_matrix_cluster}
\title{Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_P_matrix_cluster(n0, d)
}
\value{
P matrix
}
\description{
Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x
}
