% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_qp.R
\name{create_P_matrix_surv}
\alias{create_P_matrix_surv}
\title{Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_P_matrix_surv(n, trt, B_X)
}
\arguments{
\item{B_X}{n x k basis matrix of covariates}
}
\value{
P matrix
}
\description{
Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x
}
