% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment.R
\name{create_P_matrix_treatment}
\alias{create_P_matrix_treatment}
\title{Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_P_matrix_treatment(n, aux_dim)
}
\arguments{
\item{n}{Total number of units}

\item{aux_dim}{Dimension of auxiliary weights}
}
\value{
P matrix
}
\description{
Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x
}
