% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment_kernel.R
\name{create_P_matrix_treatment_kernel}
\alias{create_P_matrix_treatment_kernel}
\title{Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_P_matrix_treatment_kernel(
  n,
  X0s,
  Xtaus,
  kernel0,
  kerneltau,
  pro_trt,
  pro_ctr,
  S_factor,
  gc
)
}
\arguments{
\item{n}{Total number of units}

\item{X0s}{List of J n x d0 matrices of covariates split by group}

\item{Xtaus}{List of J n x dtau matrices of covariates split by group}

\item{kernel0}{Kernel to use for average control potential outcome function}

\item{kerneltau}{Kernel to use for average treatment effect function}

\item{pro_trt}{Vector of treatment propensity multipliers}

\item{pro_ctr}{Vector of control propensity multipliers}

\item{S_factor}{Vector of site labels}

\item{gc}{boolean indicating whether to garbage collect between operations}
}
\value{
P matrix
}
\description{
Create the P matrix for an QP that solves min_x 0.5 * x'Px + q'x
}
