% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_weights.R
\name{create_q_vector_cluster}
\alias{create_q_vector_cluster}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_cluster(ind_covs, clus_covs, trt)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{clus_covs}{n x d2 matrix of covariates for clusters}

\item{trt}{n vector of treatment assignment}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
