% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_qp.R
\name{create_q_vector_multi}
\alias{create_q_vector_multi}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_multi(Xz, trtz)
}
\arguments{
\item{Xz}{list of J n x d matrices of covariates split by group}

\item{target}{Vector of population means to re-weight to}

\item{aux_dim}{Dimension of auxiliary weights}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
