% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_kernel_QP.R
\name{create_q_vector_multi_kern}
\alias{create_q_vector_multi_kern}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_multi_kern(kern_mats, trtz)
}
\arguments{
\item{target}{Vector of population means to re-weight to}

\item{aux_dim}{Dimension of auxiliary weights}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
