% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mundlak.R
\name{create_q_vector_mundlak}
\alias{create_q_vector_mundlak}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_mundlak(ind_covs, interact_covs, trt)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{interact_covs}{n x d2 matrix of interactions between individual and cluster covariates}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
