% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_qp.R
\name{create_q_vector_surv}
\alias{create_q_vector_surv}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_surv(n, trt, B_X, Bbar_X)
}
\arguments{
\item{B_X}{n x k basis matrix of covariates}

\item{Bbar_X}{Vector of population means to re-weight to}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
