% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment.R
\name{create_q_vector_treatment}
\alias{create_q_vector_treatment}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_treatment(Xtaus, pro_trt_split, target, aux_dim)
}
\arguments{
\item{Xtaus}{List of J n x dtau matrices of covariates split by group}

\item{pro_trt_split}{List of J vectors of treatment propensity multipliers}

\item{target}{Vector of population means to re-weight to}

\item{aux_dim}{Dimension of auxiliary weights}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
