% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment_kernel.R
\name{create_q_vector_treatment_kernel}
\alias{create_q_vector_treatment_kernel}
\title{Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x}
\usage{
create_q_vector_treatment_kernel(Xtaus, pro_trt_split, Xtarget, kernel)
}
\arguments{
\item{Xtaus}{List of J n x dtau matrices of covariates split by group}

\item{pro_trt_split}{List of J vectors of treatment propensity multipliers}

\item{Xtarget}{ntarget x dtau matrix constituting the target population}

\item{kernel}{Kernel to use for average treatment effect function}
}
\value{
q vector
}
\description{
Create the q vector for an QP that solves min_x 0.5 * x'Px + q'x
}
