% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_kernel_QP.R
\name{create_scaled_I_matrix}
\alias{create_scaled_I_matrix}
\title{Create diagonal regularization matrix}
\usage{
create_scaled_I_matrix(Xz, scale_sample_size)
}
\arguments{
\item{Xz}{list of J n x d matrices of covariates split by group}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default T}

\item{n}{Total number of units}

\item{aux_dim}{Dimension of auxiliary weights}
}
\description{
Create diagonal regularization matrix
}
