% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceR.R
\name{CreateBalanceSheet}
\alias{CreateBalanceSheet}
\title{CreateBalanceTable}
\usage{
CreateBalanceSheet(journal, which = "small")
}
\arguments{
\item{journal}{The a data frame as returned from the \code{register} function
of the package \code{ledger}.}

\item{which}{Character of length one, indicating up to which hierarchy the balance
sheet has to be created. "verysmall" for 'Kleinstkapitalgesellschaften', "small" for
'Kleinkapitalgesellschaften' and "normal" for all other sizes. This will affect, how
detailed the resulting balance sheet will be.}
}
\value{
An xtable object containing the balance sheet.
}
\description{
Create a balance table as tex from a ledger journal imported using the
register function of the R package \code{ledger}.
}
\details{
The function CreateBalanceTable creates a TeX table from a ledger journal
including all the accounts required for reports according to the German HGB. Currently,
the functions result should only be used in landscape mode (for instance by including
the output in a rotatebox environment). Several other things are hard coded as well
and will be made available as an argument in the future.
}
\author{
Christoph Rust
}
