% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_qp.R
\name{estimate_se_multi}
\alias{estimate_se_multi}
\title{Estimate standard error per group}
\usage{
estimate_se_multi(y, trt, Z, weights, global_weights)
}
\arguments{
\item{y}{Vector of outcomes}

\item{trt}{Vector of treatment assignments}

\item{Z}{Vector of group indicators with J levels}

\item{weights}{Estimated primal weights as an n x J matrix}
}
\value{
att Vector of ATT estimates by group
}
\description{
Estimate standard error per group
}
