% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{get_uniform_weights}
\alias{get_uniform_weights}
\title{Get a set of uniform weights for initialization}
\usage{
get_uniform_weights(Xz)
}
\arguments{
\item{Xz}{list of J n x d matrices of covariates split by group}
}
\description{
Get a set of uniform weights for initialization
}
