% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment.R
\name{get_uniform_weights_treatment}
\alias{get_uniform_weights_treatment}
\title{Get a set of uniform weights for initialization}
\usage{
get_uniform_weights_treatment(X0s, Xtaus)
}
\arguments{
\item{X0s}{List of J n x d0 matrices of covariates split by group}

\item{Xtaus}{List of J n x dtau matrices of covariates split by group}
}
\description{
Get a set of uniform weights for initialization
}
