% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment_kernel.R
\name{get_uniform_weights_treatment_kernel}
\alias{get_uniform_weights_treatment_kernel}
\title{Get a set of uniform weights for initialization}
\usage{
get_uniform_weights_treatment_kernel(nj)
}
\arguments{
\item{nj}{Vector of sample sizes in sites}
}
\description{
Get a set of uniform weights for initialization
}
