% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_old.R
\name{l1_grp_prox}
\alias{l1_grp_prox}
\title{Group L1 norm prox operator, soft thresholding}
\usage{
l1_grp_prox(x, lam)
}
\arguments{
\item{x}{input}

\item{lam}{scaling function}
}
\value{
result of prox
}
\description{
Group L1 norm prox operator, soft thresholding
}
