% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{l1_prox}
\alias{l1_prox}
\title{L1 norm prox operator, soft thresholding}
\usage{
l1_prox(x, lam)
}
\arguments{
\item{x}{input}

\item{lam}{scaling function}
}
\value{
result of prox
}
\description{
L1 norm prox operator, soft thresholding
}
