% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_l2.R
\name{l2_balance_internal}
\alias{l2_balance_internal}
\title{Re-weight data to a target with local and global constraints}
\usage{
l2_balance_internal(
  Xz,
  targetz,
  lambda = 0,
  lowlim = 0,
  uplim = 1,
  scale_sample_size = T,
  exact_global = T,
  target_propz = NULL,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{Xz}{length J list of n_j x d matrix of covariates for each group j}

\item{targetz}{length J list of d-dimensional vectors of targets for each group j}

\item{lambda}{Regularization hyper parameter, default 0}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default T}

\item{exact_global}{Whether to enforce exact balance for overall population}

\item{target_propz}{J-dimensional vector of group proportions in the target population, must not be NULL if exact_global is TRUE}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
vector of weights solving balancing optimization problem
}
\description{
Re-weight data to a target with local and global constraints
}
