% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_old.R
\name{linf_prox}
\alias{linf_prox}
\title{L infinity norm prox operator
Uses algorithm for projection onto the L1 ball from Duchi (2008)}
\usage{
linf_prox(x, lam)
}
\arguments{
\item{x}{input}

\item{lam}{scaling function}
}
\value{
result of prox
}
\description{
L infinity norm prox operator
Uses algorithm for projection onto the L1 ball from Duchi (2008)
}
