% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceR.R
\name{MakeBalanceAccounts}
\alias{MakeBalanceAccounts}
\title{MakeBalanceAccounts}
\usage{
MakeBalanceAccounts(x, which = "small")
}
\description{
Create the account structure required by Balance sheets according to
German HGB for so-called Kaptitalgesellschaften.
}
