% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{map_to_param}
\alias{map_to_param}
\title{Map string choices to the proper parameters for the balancer sub-functions}
\usage{
map_to_param(link = c("logit", "linear", "pos-linear", "pos-enet",
  "posenet"), regularizer = c(NULL, "l1", "grpl1", "l2", "ridge", "linf",
  "nuc", "l1_all", "l1_nuc"), ipw_weights = NULL, normalized = F)
}
\arguments{
\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{normalized}{Whether to normalize the weights}

\item{type}{Find balancing weights for ATT, subgroup ATTs,
subgroup ATTs with multilevel p-score, multilevel observational studies,
ATT with missing outcomes, and heterogeneous effects}
}
\value{
Parameters for balancer
}
\description{
Map string choices to the proper parameters for the balancer sub-functions
}
