% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_weights.R
\name{maxsubset_weights}
\alias{maxsubset_weights}
\title{Find maximum effective sample size balanced set}
\usage{
maxsubset_weights(
  X,
  trt,
  lambda = 0,
  lowlim = 0,
  uplim = Inf,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment assignments}

\item{lambda}{Regularization hyper parameter, default 0}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated weights as a length n vector}
         \item{imbalance }{Imbalance in covariates as a d X J matrix}
         \item{global_imbalance}{Overall imbalance in covariates, as a length d vector }}
}
\description{
Find maximum effective sample size balanced set
}
