% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_weights.R
\name{maxsubset_weights_cluster}
\alias{maxsubset_weights_cluster}
\title{Find maximum effective sample size balanced set}
\usage{
maxsubset_weights_cluster(
  ind_covs,
  clus_covs,
  trt,
  clusters,
  lambda = 0,
  icc = 0,
  lowlim = 0,
  uplim = 1,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{ind_covs}{n x d1 matrix of covariates for individual units}

\item{clus_covs}{n x d2 matrix of covariates for clusters}

\item{trt}{Vector of treatment assignments}

\item{clusters}{n vector of cluster assignments}

\item{lambda}{Regularization hyper parameter, default 0}

\item{icc}{Intraclass correlation coefficient for regularization}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1 * number of treated/control units}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated weights as a length n vector}
         \item{imbalance }{Imbalance in covariates as a d X J matrix}
         \item{global_imbalance}{Overall imbalance in covariates, as a length d vector }}
}
\description{
Find maximum effective sample size balanced set
}
