% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_ate_qp.R
\name{multilevel_ate_qp}
\alias{multilevel_ate_qp}
\title{Re-weight treated and control sub-groups to sub-group means}
\usage{
multilevel_ate_qp(
  X,
  trt,
  Z,
  lambda = 0,
  lowlim = 0,
  uplim = 1,
  scale_sample_size = T,
  exact_global = T,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment assignments}

\item{Z}{Vector of group indicators with J levels}

\item{lambda}{Regularization hyper parameter, default 0}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default T}

\item{exact_global}{Whether to enforce exact balance for overall population}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated weights as a length n vector}
         \item{imbalance }{Imbalance in covariates as a d X J matrix}
         \item{global_imbalance}{Overall imbalance in covariates, as a length d vector }}
}
\description{
Re-weight treated and control sub-groups to sub-group means
}
