% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{nuc_prox}
\alias{nuc_prox}
\title{Nuclear norm prox operator, soft thresholding}
\usage{
nuc_prox(x, lam)
}
\arguments{
\item{x}{input}

\item{lam}{scaling function}
}
\value{
result of prox
}
\description{
Nuclear norm prox operator, soft thresholding
}
