% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_enet}
\alias{prox_enet}
\title{Prox for elastic net}
\usage{
prox_enet(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Column soft thresholded X
}
\description{
Prox for elastic net
}
