% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_enet_normalized}
\alias{prox_enet_normalized}
\title{Prox for elastic net ignoring intercept}
\usage{
prox_enet_normalized(x, lam, opts)
}
\arguments{
\item{x}{Input matrix ([intercept, coefs])}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
soft thresholded U + group-thresholded V
}
\description{
Prox for elastic net ignoring intercept
}
