% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l1}
\alias{prox_l1}
\title{L1 Prox}
\usage{
prox_l1(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Soft thresholded X
}
\description{
L1 Prox
}
