% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l1_grp}
\alias{prox_l1_grp}
\title{Group L1 Prox}
\usage{
prox_l1_grp(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Group soft thresholded X
}
\description{
Group L1 Prox
}
