% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l1_grp_l1}
\alias{prox_l1_grp_l1}
\title{Group L1 + L1 Prox}
\usage{
prox_l1_grp_l1(x, lam, opts)
}
\arguments{
\item{x}{Input matrix (two sets of parameters x = U + V)}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
soft thresholded U + group-thresholded V
}
\description{
Group L1 + L1 Prox
}
