% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l1_normalized}
\alias{prox_l1_normalized}
\title{L1 Prox ignoring intercept}
\usage{
prox_l1_normalized(x, lam, opts)
}
\arguments{
\item{x}{Input matrix ([intercept, coefs])}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
soft thresholded U + group-thresholded V
}
\description{
L1 Prox ignoring intercept
}
