% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l2}
\alias{prox_l2}
\title{L2 Prox}
\usage{
prox_l2(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Column soft thresholded X
}
\description{
L2 Prox
}
