% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l2_sq_normalized}
\alias{prox_l2_sq_normalized}
\title{Squared L2 Prox ignoring intercept}
\usage{
prox_l2_sq_normalized(x, lam, opts)
}
\arguments{
\item{x}{Input matrix ([intercept, coefs])}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
soft thresholded U + group-thresholded V
}
\description{
Squared L2 Prox ignoring intercept
}
