% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_l2_sq_Q}
\alias{prox_l2_sq_Q}
\title{Prox for 1/2 x'Qx}
\usage{
prox_l2_sq_Q(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Column soft thresholded X
}
\description{
Prox for 1/2 x'Qx
}
