% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_multilevel_ridge_nuc_normalized}
\alias{prox_multilevel_ridge_nuc_normalized}
\title{Squared L2 Prox for multilevel model, separate for global/local params + intercepts}
\usage{
prox_multilevel_ridge_nuc_normalized(x, lam, opts)
}
\arguments{
\item{x}{Input matrix (contains global and local parameters}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["alpha"] holds the ratio between global and local balance}
}
\value{
L2 squared prox values
}
\description{
Squared L2 Prox for multilevel model, separate for global/local params + intercepts
}
