% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_nuc_l1}
\alias{prox_nuc_l1}
\title{Nuclear norm + L1 Prox}
\usage{
prox_nuc_l1(x, lam, opts)
}
\arguments{
\item{x}{Input matrix (two sets of parameters x = U + V)}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
svd soft thresholded U + soft-thresholded V
}
\description{
Nuclear norm + L1 Prox
}
