% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_nuc_normalized}
\alias{prox_nuc_normalized}
\title{Nuclear norm prox ignoring intercept}
\usage{
prox_nuc_normalized(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Singular value soft thresholded X
}
\description{
Nuclear norm prox ignoring intercept
}
