% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_old.R
\name{ridge_prox}
\alias{ridge_prox}
\title{Prox for lam/2 ||x||_2^2}
\usage{
ridge_prox(x, lam)
}
\arguments{
\item{x}{input}

\item{lam}{scaling function}
}
\value{
result of prox
}
\description{
Prox for lam/2 ||x||_2^2
}
