% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_old.R
\name{softmax}
\alias{softmax}
\title{Compute numerically stable softmax with natural param eta
Logit link}
\usage{
softmax(eta)
}
\description{
Compute numerically stable softmax with natural param eta
Logit link
}
