% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softmax_weights_ipw}
\alias{softmax_weights_ipw}
\title{normalized logit weights, numerically stable}
\usage{
softmax_weights_ipw(Xc, theta, q)
}
\description{
normalized logit weights, numerically stable
}
