% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Re-weight groups to target population means}
\usage{
standardize(X, target, Z = NULL, link = c("logit", "linear",
  "pos-linear", "pos-enet", "posenet"), regularizer = c("ridge", "l1",
  "grpl1", "l2", "linf", "nuc", "l1_all", "l1_nuc"), lambda = NULL,
  nlambda = 20, lambda.min.ratio = 0.001, normalized = TRUE,
  alpha = 1, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{target}{Vector of population moments to weight towards}

\item{Z}{Vector of site indicators}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{lambda}{Regularization hyperparameter}

\item{nlambda}{Number of hyperparameters to consider}

\item{lambda.min.ratio}{Smallest value of hyperparam to consider, as proportion of smallest
value that gives the reference weights}

\item{normalized}{Whether to normalize the weights}

\item{alpha}{Elastic net parameter \eqn{\frac{1-\alpha}{2}\|\beta\|_2^2 + \alpha\|\beta\|_1}, defaults to 1}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}
  \item{alpha }{Elastic net parameter}}}

\item{interact}{Whether to interact group and individual level covariates}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Re-weight groups to target population means
}
