% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_}
\alias{standardize_}
\title{Internal function to reweight to target population means}
\usage{
standardize_(X, target, Z, weightfunc, weightfunc_ptr, proxfunc,
  balancefunc, lambda, nlambda, lambda.min.ratio, ipw_weights,
  opts = list(), prox_opts)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{target}{Vector of population means to re-weight to}

\item{Z}{Vector of hierarchical factor indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function}

\item{balancefunc}{Balance criterion measure}

\item{lambda}{Regularization hyper parameter}

\item{nlambda}{Number of hyperparameters to consider}

\item{lambda.min.ratio}{Smallest value of hyperparam to consider, as proportion of smallest
value that gives the reference weights}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}

\item{prox_opts}{List of additional arguments for prox}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Internal function to reweight to target population means
}
