% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_indirect_z}
\alias{standardize_indirect_z}
\title{Re-weight population to group z's target}
\usage{
standardize_indirect_z(
  focal_z,
  X,
  Z,
  lambda = 0,
  lowlim = 0,
  uplim = 1,
  scale_sample_size = F,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{focal_z}{Group to use as target}

\item{X}{n x d matrix of covariates}

\item{Z}{Vector of group indicators with J levels}

\item{lambda}{Regularization hyper parameter, default 0}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default T}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated primal weights as an n x J matrix}
         \item{imbalance }{Imbalance in covariates as a d X J matrix}
         }
}
\description{
Re-weight population to group z's target
}
