% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_treatment_hybrid.R
\name{standardize_treatment_hybrid}
\alias{standardize_treatment_hybrid}
\title{Re-weight groups to target population means}
\usage{
standardize_treatment_hybrid(
  X0,
  Xtau,
  target,
  S,
  Z,
  pscores,
  kernel0 = kernlab::vanilladot(),
  lambda = 0,
  lowlim = 0,
  uplim = nrow(X0),
  scale_sample_size = F,
  data_in = NULL,
  verbose = TRUE,
  return_program = TRUE,
  init_uniform = F,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  gc,
  ...
)
}
\arguments{
\item{X0}{n x d0 matrix of untransformed covariates defining the mean control response function}

\item{Xtau}{n x dtau matrix of transformed covariates defining the mean treatment effect function}

\item{target}{Vector of population means to re-weight to}

\item{S}{Numeric vector of site indicators with J levels}

\item{Z}{Numeric vector of treatment indicators with 2 levels}

\item{pscores}{Numeric vector of propensity scores}

\item{kernel0}{Kernel for control outcome covariates, default is the inner product}

\item{lambda}{Regularization hyper parameter, default 0}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default nrow(X0)}

\item{scale_sample_size}{Whether to scale the dispersion penalty by the sample size of each group, default F}

\item{data_in}{Optional list containing pre-computed objective matrix/vector and constraints (without regularization term)}

\item{verbose}{Whether to show messages, default T}

\item{return_program}{Whether to return the objective matrix and vector and constraints, default T}

\item{init_uniform}{Wheter to initialize solver with uniform weights, default F}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{gc}{boolean indicating whether to garbage collect between operations}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated primal weights as an n x J matrix}
         \item{imbalance_0 }{Imbalance in covariates defining mean control response function as a d0 x J matrix}
         \item{imbalance_tau }{Imbalance in covariates defining mean treatment effect function as a dtau x J matrix}
         \item{data_out }{List containing elements of QP min 0.5 x'Px + q'x st l <= Ax <= u \itemize{
                 \item{P, q}{}
                 \item{constraints }{A, l , u}
}}}
}
\description{
Re-weight groups to target population means
}
