% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_intervention.R
\name{stochastic_int}
\alias{stochastic_int}
\title{Balance towards a stochastic intervention}
\usage{
stochastic_int(
  X,
  trt,
  stoch_int,
  Z = NULL,
  lambda = 0,
  lowlim = 0,
  uplim = 1,
  exact_global = FALSE,
  verbose = TRUE,
  eps_abs = 1e-05,
  eps_rel = 1e-05,
  ...
)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of K-level treatment assignments}

\item{stoch_int}{n x K matrix of treatment probabilities under stochastic intervention}

\item{Z}{Vector of group indicators with J levels}

\item{lambda}{Regularization hyper parameter, default 0}

\item{lowlim}{Lower limit on weights, default 0}

\item{uplim}{Upper limit on weights, default 1}

\item{exact_global}{Whether to enforce exact balance for overall population}

\item{verbose}{Whether to show messages, default T}

\item{eps_abs}{Absolute error tolerance for solver}

\item{eps_rel}{Relative error tolerance for solver}

\item{...}{Extra arguments for osqp solver}
}
\value{
\itemize{
         \item{weights }{Estimated weights as a length n vector}
         \item{imbalance }{Imbalance in covariates as a d X J matrix}
         \item{global_imbalance}{Overall imbalance in covariates, as a length d vector }}
}
\description{
Balance towards a stochastic intervention
}
