% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancer.R, R/main_funcs.R
\docType{package}
\name{balancer}
\alias{balancer}
\title{balancer: A package for treatment effect heterogeneity and missing data in observational studies through matrix-constrained balancing weights}
\usage{
balancer(
  X,
  trt,
  y = NULL,
  Z = NULL,
  V = NULL,
  type = c("att", "subgrp", "subgrp_multi"),
  link = c("logit", "linear", "pos-linear", "pos-enet", "posenet"),
  regularizer = c(NULL, "l1", "grpl1", "l2", "ridge", "linf", "nuc", "l1_all",
    "l1_nuc"),
  lambda = NULL,
  nlambda = 20,
  lambda.min.ratio = 0.001,
  interact = F,
  normalized = TRUE,
  alpha = 1,
  Q = NULL,
  ipw_weights = NULL,
  mu0 = NULL,
  opts = list()
)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{y}{Vector of outcomes to estimate effect(s). If NULL then only return weights}

\item{Z}{Vector of subgroup indicators or observed indicators}

\item{V}{Group level covariates}

\item{type}{Find balancing weights for ATT, subgroup ATTs,
subgroup ATTs with multilevel p-score, multilevel observational studies,
ATT with missing outcomes, and heterogeneous effects}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{lambda}{Regularization hyperparameter}

\item{nlambda}{Number of hyperparameters to consider}

\item{lambda.min.ratio}{Smallest value of hyperparam to consider, as proportion of smallest
value that gives the reference weights}

\item{interact}{Whether to interact group and individual level covariates}

\item{normalized}{Whether to normalize the weights}

\item{alpha}{Elastic net parameter \eqn{\frac{1-\alpha}{2}\|\beta\|_2^2 + \alpha\|\beta\|_1}, defaults to 1}

\item{Q}{Matrix for generalized ridge, if null then use inverse covariance matrix}

\item{ipw_weights}{Separately estimated IPW weights to measure dispersion against, default is NULL}

\item{mu0}{Optional estimates of the potential outcome under control, default is NULL}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}
  \item{alpha }{Elastic net parameter}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
