% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancer.R
\docType{package}
\name{balancer}
\alias{balancer}
\alias{balancer-package}
\title{balancer: A package for treatment effect heterogeneity and missing data in observational studies through matrix-constrained balancing weights}
\description{
balancer: A package for treatment effect heterogeneity and missing data in observational studies through matrix-constrained balancing weights
}
