% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{balancer_multi}
\alias{balancer_multi}
\title{Balancing weights for ATT (with hierarchical structure)}
\usage{
balancer_multi(X, V = NULL, trt, Z, weightfunc, weightfunc_ptr, proxfunc,
  hyperparam, ridge = F, interact = F, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{V}{n x k matrix of group level covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of hierarchical factor indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{hyperparam}{Regularization hyper parameter}

\item{ridge}{Whether to use ridge penalty in dual}

\item{interact}{Whether to interact group and individual level characterisitics}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for ATT (with hierarchical structure)
}
