% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{balancer_subgrp}
\alias{balancer_subgrp}
\title{Balancing weights for ATT (in subgroups)}
\usage{
balancer_subgrp(X, trt, Z = NULL, weightfunc, weightfunc_ptr, proxfunc,
  hyperparam, ridge, Q = NULL, kernel, kern_param = 1, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of subgroup indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function, or "ridge"}

\item{hyperparam}{Regularization hyper parameter}

\item{ridge}{Whether to use L2 penalty in dual}

\item{Q}{m x m matrix to tie together ridge penalty, default: NULL,
if TRUE, use covariance of treated groups}

\item{kernel}{What kernel to use, default NULL}

\item{kern_param}{Hyperparameter for kernel}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for ATT (in subgroups)
}
