% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{balancer_subgrp}
\alias{balancer_subgrp}
\title{Balancing weights for ATT (in subgroups)}
\usage{
balancer_subgrp(X, trt, Z = NULL, weightfunc, weightfunc_ptr, proxfunc,
  hyperparam, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{Z}{Vector of subgroup indicators}

\item{weightfunc}{Derivative of convex conjugate of dispersion function (possibly normalized)}

\item{weightfunc_ptr}{Pointer to weightfunc}

\item{proxfunc}{Prox operator of regularization function}

\item{hyperparam}{Regularization hyper parameter}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Balancing weights for ATT (in subgroups)
}
