% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{create_constraints}
\alias{create_constraints}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints(X, target, Z, lowlim, uplim)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{target}{Vector of population means to re-weight to}

\item{Z}{Vector of group indicators}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
