% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_qp.R
\name{create_constraints_multi}
\alias{create_constraints_multi}
\title{Create the constraints for QP: l <= Ax <= u}
\usage{
create_constraints_multi(Xz, trtz, lowlim, uplim, exact_global, verbose)
}
\arguments{
\item{Xz}{list of J n x d matrices of covariates split by group}

\item{lowlim}{Lower limit on weights}

\item{uplim}{Upper limit on weights}

\item{target}{Vector of population means to re-weight to}
}
\value{
A, l, and u
}
\description{
Create the constraints for QP: l <= Ax <= u
}
